Fixed memory leak
It appears that tkinter objects need to be destroyed manually.
So now I know that.
All buttons now stack in columns by 5. Added 'exit' button.
Redisigned Fleet and Grid classes. Now the first one does not need the
second one to be initialized.
UI buttons no longer have lambda functions as their commands.
NN now thinks that if there is a ship on the top cell, then there is now
ship on the current cell. What an idiot lol.
I should add a cheker to see if the program able to place all the ships
in the current grid.